% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_DB_info}
\alias{get_DB_info}
\title{Get database info}
\usage{
get_DB_info(connectionString)
}
\arguments{
\item{connectionString}{Connection string to SQL server}
}
\value{
Returns data.frame and data.table
}
\description{
This function retrieves basic info about database defined
in SQL Server connection string.
}
\note{
How to set up SQL Server connection string see \link{set_connString}.
}
\examples{
get_DB_info(connectionString)
}
