% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_table_info}
\alias{get_table_info}
\title{Get table info}
\usage{
get_table_info(connectionString, sqltabname)
}
\arguments{
\item{connectionString}{Connection string to SQL server}

\item{sqltabname}{SQL table name}
}
\value{
Returns data.frame and data.table
}
\description{
This function retrieves basic info about table
}
\note{
How to set up SQL Server connection string see \link{set_connString}.
}
\examples{
get_table_info(connectionString, "dbo.tableName")
}
