% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pull_data}
\alias{pull_data}
\title{Pull data from SQL server via SQL query}
\usage{
pull_data(connectionString, sqltask, showprogress = FALSE)
}
\arguments{
\item{connectionString}{Connection string to SQL server}

\item{sqltask}{SQL query for selecting data on SQL server}

\item{showprogress}{Showing progress (default value is FALSE)}
}
\value{
Returns data.frame and data.table
}
\description{
This function pulls the data from SQL server
via SQL query and returns data.table and data.frame object
}
\note{
How to set up SQL Server connection string see \link{set_connString}.
}
\examples{
\dontrun{
pull_data(connectionString, "SELECT * FROM dbo.TableName")
}
}
