% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{send_SQL_task}
\alias{send_SQL_task}
\title{Drop, delete or create table}
\usage{
send_SQL_task(connectionString, sqltask)
}
\arguments{
\item{connectionString}{Connection string to SQL server}

\item{sqltask}{SQL query for retrieving data from SQL server}
}
\description{
This function allows user to drop table, delete rows in table
or create new table on SQL Server
}
\note{
How to set up SQL Server connection string see \link{set_connString}.
}
\examples{
send_SQL_task(connString, "CREATE TABLE dbo.TableName (ID int not null, Name varchar(100))")
send_SQL_task(connString, "DELETE FROM dbo.TableName WHERE ColumnName = 'SomeValue'")
send_SQL_task(connString, "DROP TABLE dbo.TableName")
}
