% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_connString}
\alias{set_connString}
\title{Create connection string from function parameters}
\usage{
set_connString(datasource, database, usr, pwd)
}
\arguments{
\item{datasource}{Server name}

\item{database}{Database name}

\item{usr}{Username}

\item{pwd}{Password}
}
\value{
Connection string
}
\description{
This function defines SQL Server connection string
from function parameters.
}
\note{
If username and password missing or empty \strong{Integrated Security=True} is used in connection string instead.
}
\examples{
set_connString("LAPTOP-USER\\\\SQLEXPRESS", "Database_Name")
}
